new worker for embroidery turtle + improvements
This commit prepares the transition from the lazer turtle tool
to a more generic "crafty turtle" tool.

The basic idea of the embroidery turtle is to draw alternately
on the front and on the back side of the plane, very much like
embroidery thread works its way through the fabric.

The turtle worker creates 3 SVG paths at the same time

- the default turtle path
- the path of stitches as it appears on the front side
- the path od stitches as it appears on the reverse side

Additional improvements

- no more hardcoded locations
- variables instantiated from html form state
- menu to switch between SVG modes